TableReaderSite/requirements.txt
Django~=5.0.6
pandas~=2.2.2
djangorestframework~=3.15.2
toml~=0.10.2
requests~=2.32.3